Made Sprite class a base class for use in other sprites
Error with template, will fix in next commit
