+++ HardestGame/HardestGame.vcxproj
@@ -112,13 +112,13 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>_DEBUG;_CONSOLE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>C:\Users\Ethan\Desktop\Spring2025\SFML-3.0.0\include;C:\Users\ajayr\OneDrive\Documents\CPTS 122\SFML-3.0.0\include;C:\Users\egreg\source\repos\SFML-3.0.0\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>C:\Users\schoolmason\Desktop\spring 122\SFML-3.0.0\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp17</LanguageStandard>
     </ClCompile>
     <Link>
       <SubSystem>Console</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalLibraryDirectories>C:\Users\Ethan\Desktop\Spring2025\SFML-3.0.0\lib;C:\Users\ajayr\OneDrive\Documents\CPTS 122\SFML-3.0.0\lib;C:\Users\egreg\source\repos\SFML-3.0.0\lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
+      <AdditionalLibraryDirectories>C:\Users\schoolmason\Desktop\spring 122\SFML-3.0.0\lib;C:\Users\Ethan\Desktop\Spring2025\SFML-3.0.0\lib;C:\Users\ajayr\OneDrive\Documents\CPTS 122\SFML-3.0.0\lib;C:\Users\egreg\source\repos\SFML-3.0.0\lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
       <AdditionalDependencies>sfml-system-d.lib;sfml-window-d.lib;sfml-graphics-d.lib;sfml-audio-d.lib</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
@@ -130,7 +130,7 @@
       <SDLCheck>true</SDLCheck>
       <PreprocessorDefinitions>NDEBUG;_CONSOLE;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <ConformanceMode>true</ConformanceMode>
-      <AdditionalIncludeDirectories>C:\Users\Ethan\Desktop\Spring2025\SFML-3.0.0\include;C:\Users\ajayr\OneDrive\Documents\CPTS 122\SFML-3.0.0\include;C:\Users\egreg\source\repos\SFML-3.0.0\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>C:\Users\schoolmason\Desktop\spring 122\SFML-3.0.0\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <LanguageStandard>stdcpp17</LanguageStandard>
     </ClCompile>
     <Link>
@@ -138,7 +138,7 @@
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalLibraryDirectories>C:\Users\Ethan\Desktop\Spring2025\SFML-3.0.0\lib;C:\Users\ajayr\OneDrive\Documents\CPTS 122\SFML-3.0.0\lib;C:\Users\egreg\source\repos\SFML-3.0.0\lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
+      <AdditionalLibraryDirectories>C:\Users\schoolmason\Desktop\spring 122\SFML-3.0.0\lib;C:\Users\Ethan\Desktop\Spring2025\SFML-3.0.0\lib;C:\Users\ajayr\OneDrive\Documents\CPTS 122\SFML-3.0.0\lib;C:\Users\egreg\source\repos\SFML-3.0.0\lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
       <AdditionalDependencies>sfml-system.lib;sfml-window.lib;sfml-graphics.lib;sfml-audio.lib</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
@@ -148,6 +148,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Level.hpp" />
+    <ClInclude Include="Player.hpp" />
     <ClInclude Include="Sprite.hpp" />
     <ClInclude Include="Wall.hpp" />
   </ItemGroup>
