+++ HardestGame/Level.hpp
@@ -1,4 +1,5 @@
 #pragma once
 #include "Sprite.hpp"
+#include "Player.hpp"
  
-bool levelOne(sf::RenderWindow& window, Sprite* player, sf::Sprite& backgroundTexture, sf::Text& lvlMsg);
+bool levelOne(sf::RenderWindow& window, Sprite<Player>* player, sf::Sprite& backgroundTexture, sf::Text& lvlMsg);
